Make chopper streams an instrument concern
Choppers are an instrument property, not a workflow choice: how they are set up
and which streams they produce belongs to the instrument's data model, with the
wavelength-LUT workflow as the (currently sole) consumer. The instrument
already owns chopper config via the ``choppers`` and ``chopper_delay_atol``
fields; this moves the stream side to match.

Move the chopper stream-name conventions and declare_chopper_setpoint_streams
to a new config/chopper.py, and invoke the declaration automatically from
Instrument.__post_init__ for any instrument that declares choppers — before the
f144 snapshot that seeds timeseries specs. Declaring choppers now implies the
synthetic delay_setpoint streams; the per-instrument specs.py call is gone.

This also fixes the import direction (the instrument layer no longer reaches
into handlers for stream naming) and centralizes the provisional delay-stream
convention in one place. The workflow side keeps what is genuinely its own:
the Sciline context keys, the DiskChoppers provider, and the factory attach.
